+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 * Added `-R` option for specifying extra commandline parameters for emulators.
 
-* Added full 16-bit multiplication. 
+* Added full 16-bit multiplication and unsigned division. 
 
 * Added preliminary support for EasyFlash.